feature(mailjet): Inject Mailjet service and create enum type
