+++ back-end/src/common/providers/interfaces/events.interface.ts
@@ -1,3 +1,8 @@
 export enum Events {
 	poSignup = 'mailjet.signup.po',
+	userSignup = 'mailjet.signup.user',
+}
+
+export enum MaijetTemplate {
+	signup = 4422523,
 }
